+++ Diary/Model/CoreData/Diary+CoreDataClass.swift
