+++ Diary/Model/CoreData/Diary+CoreDataProperties.swift
