+++ Diary/Model/DTO/WeatherResult.swift
