+++ Diary/Model/Namespace/AlertNamespace.swift
