+++ Diary/Model/Namespace/ButtonNamespace.swift
